% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxed.labels.R
\name{boxed.labels}
\alias{boxed.labels}
\title{Place labels in boxes}
\usage{
boxed.labels(x, y = NA, labels, bg = ifelse(match(par("bg"),
  "transparent", 0), "white", par("bg")), border = TRUE, xpad = 1.2,
  ypad = 1.2, srt = 0, cex = 1, adj = 0.5, xlog = FALSE,
  ylog = FALSE, ...)
}
\arguments{
\item{x, y}{x and y position of the centers of the labels. x can be an \code{\link{xy.coords}} list.}

\item{labels}{Text strings}

\item{bg}{The fill color of the rectangles on which the labels are displayed (see Details).}

\item{border}{Whether to draw borders around the rectangles.}

\item{xpad, ypad}{The proportion of the rectangles to the extent of the text within.}

\item{cex}{Character expansion. See \code{\link{text}}.}

\item{adj}{left/right adjustment. If this is set outside the function, the box will not be aligned properly.}

\item{xlog}{Whether the X axis is a log axis}

\item{ylog}{Whether the Y axis is a log axis}

\item{...}{additional arguments passed to \code{\link{text}}.}
}
\value{
nil
}
\description{
This function is a workaround for an issue with \code{\link[plotrix]{boxed.labels}}, where adj doesn't seem to function properly.
}
\details{
The label(s) are displayed on a rectangular background. This may be useful for visibility and is the reason that "transparent" background is not available. With the default textcol=NA, the function tries to work out whether white or black text will be more easily read based on the background color and displays the text accordingly. If the user specifies text colors in the additional arguments, these colors will override the automatic white/black above - see the last example. \cr
}
