% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_difference.R
\name{calculate_difference}
\alias{calculate_difference}
\title{Calculate Difference}
\usage{
calculate_difference(data1, data2)
}
\arguments{
\item{data1}{The 1st data object}

\item{data2}{The 2nd data object}
}
\value{
A data object representing the difference
}
\description{
This function calculates the difference between two data objects
}
\examples{
calculate_difference(data1,data2)
}
