% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.legend.R
\name{color.legend.jms}
\alias{color.legend.jms}
\title{Colour Data}
\usage{
color.legend.jms(
  xl,
  yb,
  xr,
  yt,
  legend,
  rect.col,
  cex = 1,
  align = "lt",
  gradient = "x",
  ...
)
}
\description{
This function is a copy of \code{\link[plotrix]{color.scale}} but without the override of par('xpd').
}
