% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.scale.R
\name{color.scale.jms}
\alias{color.scale.jms}
\title{Colour Data}
\usage{
color.scale.jms(x, cs1 = c(0, 1), cs2 = c(0, 1), cs3 = c(0, 1),
  alpha = 1, extremes = NA, na.color = NA, xrange = NULL,
  color.spec = "rgb", unsafe = FALSE)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame}

\item{cs1, cs2, cs3}{Colour parameters for scaling x}

\item{alpha}{Value for transparency in colours.}

\item{extremes}{The colours for the extreme values of x (RGB only).}

\item{na.color}{The colour to use for NA values of x.}

\item{xrange}{An explicit range to use in the transformation.}

\item{color.spec}{The color specification to use in the transformation. Anything other than "rgb", "hsv" or "hcl" will almost certainly fail.}
}
\description{
This function is a workaround for an issue with \code{\link[plotrix]{color.scale}}.
Calculating the number of colours must be performed after the check of xrange.
}
\details{
color.scale calculates a sequence of colors by a linear transformation of the numeric values supplied into the ranges for the three color parameters.If only one number is supplied for a color range, that color remains constant for all values of x. If more than two values are supplied, the x values will be split into equal ranges (one less than the number of colors) and the transformation carried out on each range. Values for a color range must be between 0 and 1 for the RGB or HSV specifications, and between 0 and 360 (cs1) and 0 to 100 (cs2 and cs3) for the HCL specifications.

IMPORTANT: If x has fewer values than the number of values in the color parameters, it will usually return incorrect colors. This is usually only a problem when using color.legend with a small number of rectangles in the legend as color.legend calls color.scale to calculate the color rectangles.

If extremes is not NA, the ranges will be calculated from its values using col2rgb, even if ranges are also supplied. extremes allows the user to just pass the extreme color values in any format that col2rgb will accept. Note that this forces the color specification to RGB.

If the user wants to specify a range of values with xrange, it must at least include the range of x values. This can be useful when there is a notional range like 0-100% that the values do not cover, or when several series of values with different ranges are to be assigned the same color scale.

The user may not want the color scheme to be continuous across some critical point, often zero. In this case, color scale can be called separately for the values below and above zero. I may get around to adding an argument to do this in one shot. Until then, see the second example for color2D.matplot and also the diverge.hcl and diverge.hsv functions in the colorspace package.
}
