% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.jms.R
\name{combine.data.objects}
\alias{combine.data.objects}
\title{Combine a list of data objects}
\usage{
combine.data.objects(objects)
}
\value{
A jms.data.object containing the data
}
\description{
This function combines jms.data.objects
}
\examples{
combine.data.objects(objects)
}
