% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.jms.data.object.R
\name{create_data_type}
\alias{create_data_type}
\title{Extend jms.data.object for a custom class}
\usage{
create_data_type(name, xlab, ylab, environment)
}
\description{
Creates the following functions:
\itemize{
\item <name>.data.object
\item as.<name>.data.object
\item is.<name>.data.object
\item [.<name>.data.object
\item read.table.name
}
}
