% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{export.jms.data.object}
\title{Export a data object}
\usage{
export(x, path)

\method{export}{jms.data.object}(x, path)
}
\arguments{
\item{x}{The data object}

\item{path}{The path to a csv file in which to save the data}
}
\description{
Export a data object
}
