% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.layout.R
\name{grid.layout}
\alias{grid.layout}
\title{Layout plots in grid form}
\usage{
grid.layout(n, show = FALSE, device = "both")
}
\arguments{
\item{n}{number of plots}

\item{show}{Show the resultant layout using \code{\link{layout.show}}?}

\item{device}{Which device to setup - one of 'base', 'interactive', or 'both'}
}
\description{
Sets up a grid layout for plots. Currently up to 9 plots are supported.
}
