% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilayout.R
\name{ilayout.addPlot}
\alias{ilayout.addPlot}
\title{Add a plot to an interactive layout}
\usage{
ilayout.addPlot(graph)
}
\arguments{
\item{graph}{The \code{\link[dygraphs]{dygraph}} to add.}
}
\description{
Add a plot to an interactive layout
}
\keyword{internal}
