% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPlot.jms.data.object.R
\name{iPlot}
\alias{iPlot}
\alias{iPlot.jms.data.object}
\title{Make an interactive plot for a data object}
\usage{
iPlot(x)

\method{iPlot}{jms.data.object}(data, xlim = NULL, ylim = NULL,
  axes = c(1, 2), xlab = xlab_(data), ylab = ylab_(data),
  col = par("col"), lwd = 1, pch = NA)
}
\arguments{
\item{x}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}

\item{xlim}{numeric vectors of length 2, giving the x and y
    coordinates ranges.}

\item{ylim}{numeric vectors of length 2, giving the x and y
    coordinates ranges.}

\item{axes}{a logical value indicating whether both axes should be drawn on
    the plot.  Use \link{graphical parameter} \code{"xaxt"} or \code{"yaxt"}
    to suppress just one of the axes.}

\item{xlab}{a label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{a label for the y axis, defaults to a description of \code{y}.}

\item{col}{color code or name, see \code{\link{colors}},
    \code{\link{palette}}.  Here \code{NULL} means colour 0.}

\item{lwd}{line width, also used for (non-filled) plot symbols, see
    \code{\link{lines}} and \code{\link{points}}.}

\item{pch}{character or integer code for kind of points, see
    \code{\link{points.default}}.}
}
\description{
Make an interactive plot for a data object
}
