% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jms.data.object.R
\name{lines.jms.data.object}
\alias{lines.jms.data.object}
\title{Draw lines for a data object}
\usage{
\method{lines}{jms.data.object}(x, offset = 1/sqrt(ncol(x) - 1),
  col = par("col"), ...)
}
\arguments{
\item{data}{The object to plot}
}
\description{
This function plots a data object
}
\examples{
lines(data)
}
