% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.jms.R
\name{load.combine}
\alias{load.combine}
\title{Combine a list of data objects}
\usage{
load.combine(objects)
}
\value{
A jms.data.object containing the data
}
\description{
This function combines jms.data.objects
}
\examples{
load.combine(objects)
}
