% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.jms.R
\name{load.jms}
\alias{load.jms}
\title{Read data files}
\usage{
load.jms(path, func, ...)
}
\arguments{
\item{path}{The path to the file / directory}

\item{func}{The function to read the file}

\item{...}{Additional parameters are passed to \code{\link[Plotting.Utils]{load.directory}} or \code{\link[Plotting.Utils]{load.file}}}
}
\value{
A jms.data.object containing the data
}
\description{
This function reads one or more data files in a directory
}
\examples{
load.jms('/path/to/directory', load_function, ext='ext')
load.jms('/path/to/file.ext', load_function)
load.jms(c('/path/to/file.ext','/path/to/file2.ext'))
}
