% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTIL-nearest.R
\name{nearest}
\alias{nearest}
\title{Find the nearest rows in a dataframe}
\usage{
nearest(data, ..., as.list = F)
}
\arguments{
\item{data}{The data frame}

\item{...}{the data to match e.g. column=c(1,2,3,4)}

\item{as.list}{Return a list even if there is only one column to search}
}
\value{
list of columns, for each column name: a data.frame of the nearest rows
        if as.list is FALSE and there is only one column the data.frame is returned
        without an encapsulating list
}
\description{
Find the nearest rows in a dataframe
}
\examples{
nearest(rock, area=c(6775, 4895, 9819, 4990))

nearest(rock, area=c(6775, 4895, 9819, 4990), shape=0.31)
}
