% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot.R
\name{new_plot}
\alias{new_plot}
\title{Creates a blank plot with the specified limits}
\usage{
new_plot(xlim, ylim)
}
\arguments{
\item{xlim, ylim}{Numeric vectors of length 2, giving the x and y coordinates ranges.}
}
\description{
Creates a blank plot with the specified limits
}
