% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jms.data.object.R
\name{plot.jms.data.object}
\alias{plot.jms.data.object}
\title{Plot a data object}
\usage{
\method{plot}{jms.data.object}(data, offset = 1/sqrt(ncol(data) - 1),
  xlim = NULL, ylim = NULL, xaxt = par("xaxt"), yaxt = par("yaxt"), ...)
}
\arguments{
\item{data}{The object to plot}
}
\description{
This function plots a data object
}
\examples{
plot(data)
}
