% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.scale.R
\name{rescale.jms}
\alias{rescale.jms}
\title{Rescale Data}
\usage{
rescale.jms(x, newrange)
}
\description{
This function is a workaround for an issue with \code{\link[plotrix]{rescale}}.
Where all values for x are the same, we should still make sure they fall within the required range. We thus return the lower limit for all values.
}

