% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pretty.R
\name{rescale}
\alias{rescale}
\title{Calculate a scale factor to transform between vectors}
\usage{
rescale(y1, y2)
}
\arguments{
\item{y1, y2}{Numeric vectors}
}
\value{
c(scale,offset) to transform y2 into the same range as y1
}
\description{
Calculate a scale factor to transform between vectors
}
