% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pretty.R
\name{tick_interval}
\alias{tick_interval}
\title{Calculates a tick inteval to make pretty plots}
\usage{
tick_interval(range, frac = FALSE)
}
\arguments{
\item{range}{The axis range}

\item{frac}{Allow fractional intervals?}
}
\description{
Calculates a tick inteval to make pretty plots
}
