% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xycols.R
\name{xcol}
\alias{xcol}
\alias{xcol<-}
\alias{ycol}
\alias{ycol<-}
\alias{y2col}
\alias{y2col<-}
\title{Get or set the x or y column(s) for a data object}
\usage{
xcol(x)

xcol(x) <- value

ycol(x)

ycol(x) <- value

y2col(x)

y2col(x) <- value
}
\arguments{
\item{x}{The data object}
}
\value{
The column(s)
}
\description{
Get or set the x or y column(s) for a data object
}
