% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_label.R
\name{add_plot_label}
\alias{add_plot_label}
\title{Add a label to a plot}
\usage{
add_plot_label(xoffset = 0, yoffset = 0, label, ...)
}
\arguments{
\item{xoffset}{Offset in inches from the left limit of the plot}

\item{yoffset}{Offset in inches from the top limit of the plot}

\item{label}{The label text}

\item{...}{Additional parameters passed to \code{\link{text}}}
}
\description{
Add a label to a plot
}

