% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurlyBrace.R
\name{CurlyBrace}
\alias{CurlyBrace}
\title{Add a horizontal curly brace to a plot}
\usage{
CurlyBrace(start, end, width, ...)
}
\arguments{
\item{start}{x coordinate of the start of the brace}

\item{end}{x coordinate of the end of the brace}

\item{width}{Width of the brace}

\item{...}{Additional parameters passed to \code{\link{lines}}}
}
\description{
Add a horizontal curly brace to a plot
}

