% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.layout.R
\name{grid.layout}
\alias{grid.layout}
\title{Layout plots in grid form}
\usage{
grid.layout(n, show = FALSE)
}
\arguments{
\item{n}{number of plots}

\item{show}{Show the resultant layout using \code{\link{layout.show}}?}
}
\description{
Layout plots in grid form
}
