% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilayout.R
\name{ilayout}
\alias{ilayout}
\alias{ilayout.addPlot}
\alias{ilayout.hideTraceFromPlot}
\alias{ilayout.show}
\alias{layout.all}
\title{Specifying Complex Plot Arrangements for Interactive Plots}
\usage{
ilayout(mat, widths = rep.int(1, ncol(mat)), heights = rep.int(1, nrow(mat)))

ilayout.addPlot(graph)

ilayout.hideTraceFromPlot(trace, plot = 1, visibility = FALSE)

ilayout.show(n)

layout.all(...)
}
\arguments{
\item{mat}{a matrix object specifying the location of the next
    \eqn{N} figures on the output device.  Each value in the matrix must
    be \code{0} or a positive integer.  If \eqn{N} is the largest
    positive integer in the matrix, then the integers
    \eqn{\{1, \dots, N-1\}}{{1, \\dots, N-1}} must also appear at least once in the matrix.}

\item{widths}{a vector of values for the widths of columns on the
    device.  Relative widths are specified with numeric values.    Absolute
    widths (in centimetres) are specified with the \code{lcm()} function (see
    examples).}

\item{heights}{a vector of values for the heights of rows on the device.
    Relative and absolute heights can be specified, see \code{widths} above.}

\item{graph}{The \code{\link[dygraphs]{dygraph}} to add.}

\item{trace}{Trace # of plot}

\item{plot}{\code{ilayout} plot number}

\item{n}{number of figures to plot.}
}
\description{
\code{ilayout} divides the device up into as many rows and columns
as there are in matrix \code{mat}, with the column-widths and the
row-heights specified in the respective arguments.
}
\details{
Figure \emph{i} is allocated a region composed from a subset of
these rows and columns, based on the rows and columns in which \emph{i}
occurs in \code{mat}. \cr
\code{ilayout.addPlot} adds a plot to the current layout \cr
\code{ilayout.show}(n) plots (part of) the current layout,
namely the outlines of the next n figures. \cr
\code{ilayout.all} Sets up both the default plot device and the interactive
plot device.
}
