% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jms.data.object.R
\name{plot.jms.data.object}
\alias{plot.jms.data.object}
\title{Plot a data object}
\usage{
\method{plot}{jms.data.object}(x, offset = 1/sqrt(ncol(x) - 1), xlim = NULL,
  ylim = NULL, xlab = xlab_(x), ylab = ylab_(x), xaxt = par("xaxt"),
  yaxt = par("yaxt"), .extend_y = c(0, 0), y_axis = 2, ...)
}
\arguments{
\item{x}{The object to plot}
}
\description{
This function plots a data object
}
\examples{
plot(data)
}
