% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_interval.R
\name{pretty_axes}
\alias{pretty_axes}
\title{Draws a bounding box and pretty axes}
\usage{
pretty_axes(min, max, x_axis = 1, y_axis = 2, frac = FALSE, div = 1,
  ensureZero = TRUE)
}
\arguments{
\item{min}{vector of minimum axis limits c(x, y)}

\item{max}{vector of maximum axis limits c(x, y)}

\item{x_axis}{Side on which to draw the x axis}

\item{y_axis}{Side on which to draw the y axis}

\item{frac}{Allow fractional intervals?}

\item{div}{Factor by which the minor tick interval is smaller than the major tick interval c(x,y)}

\item{ensureZero}{If 0 is within the range, ensure it has a tick mark}
}
\description{
Draws a bounding box and axes with tick positions generated by \code{\link{pretty_ticks}}
}
