% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot.R
\name{pretty_plot}
\alias{pretty_plot}
\title{Makes a new plot with pretty axes}
\usage{
pretty_plot(xlim, ylim, x_axis = 1, y_axis = 2, frac = FALSE, div = 2,
  flexible = TRUE, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{xlim, ylim}{Numeric vectors of length 2, giving the x and y coordinates ranges.}

\item{x_axis, y_axis}{Side on which to draw the x,y axes}

\item{frac}{Allow fractional intervals?}

\item{div}{Factor by which the minor tick interval is smaller than the major tick interval c(x,y)}

\item{flexible}{Allow min and max to be adjusted for prettier results?}

\item{xlab, ylab}{Titles for the x,y axes}

\item{...}{Additional parameters are passed to the underlying function calls}
}
\description{
Makes a new plot with \code{\link{new_plot}} and a bounding box and axesusing \code{\link{pretty_axes}}
}
