% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_interval.R
\name{pretty_ticks}
\alias{pretty_ticks}
\title{Calculates tick locations to make pretty plots}
\usage{
pretty_ticks(min, max, frac = FALSE, div = 1, flexible = TRUE)
}
\arguments{
\item{min, max}{The minimum / maximum axis limits}

\item{frac}{Allow fractional intervals?}

\item{div}{Divide the calculated interval by this number (to generate minor ticks)}

\item{flexible}{Allow min and max to be adjusted for prettier results?}
}
\description{
Calculates tick locations to make pretty plots
}
